+++ sdks/go/pkg/beam/runners/prism/internal/execute.go
@@ -37,7 +37,6 @@ import (
 	"github.com/apache/beam/sdks/v2/go/pkg/beam/runners/prism/internal/worker"
 	"golang.org/x/exp/maps"
 	"golang.org/x/sync/errgroup"
-	"google.golang.org/protobuf/encoding/prototext"
 	"google.golang.org/protobuf/proto"
 )
 
@@ -286,13 +285,25 @@ func executePipeline(ctx context.Context, wks map[string]*worker.W, j *jobservic
 					//slog.Warn("teststream bytes", "value", string(v), "bytes", v)
 					return v
 				}
-				// Hack for Java Strings in test stream, since it doesn't encode them correctly.
-				forceLP := cID == "StringUtf8Coder" || cID != pyld.GetCoderId()
+				// If the TestStream coder needs to be LP'ed or if it is a coder that has different
+				// behaviors between nested context and outer context (in Java SDK), then we must
+				// LP this coder and the TestStream data elements.
+				forceLP := cID != pyld.GetCoderId() ||
+					coders[cID].GetSpec().GetUrn() == urns.CoderStringUTF8 ||
+					coders[cID].GetSpec().GetUrn() == urns.CoderBytes ||
+					coders[cID].GetSpec().GetUrn() == urns.CoderKV
 				if forceLP {
 					// slog.Warn("recoding TestStreamValue", "cID", cID, "newUrn", coders[cID].GetSpec().GetUrn(), "payloadCoder", pyld.GetCoderId(), "oldUrn", coders[pyld.GetCoderId()].GetSpec().GetUrn())
 					// The coder needed length prefixing. For simplicity, add a length prefix to each
 					// encoded element, since we will be sending a length prefixed coder to consume
 					// this anyway. This is simpler than trying to find all the re-written coders after the fact.
+					// This also adds a LP-coder for the original coder in comps.
+					cID, err := forceLpCoder(pyld.GetCoderId(), comps.GetCoders())
+					if err != nil {
+						panic(err)
+					}
+					slog.Debug("teststream: add coder", "coderId", cID)
+
 					mayLP = func(v []byte) []byte {
 						var buf bytes.Buffer
 						if err := coder.EncodeVarInt((int64)(len(v)), &buf); err != nil {
@@ -303,6 +314,13 @@ func executePipeline(ctx context.Context, wks map[string]*worker.W, j *jobservic
 						}
 						//slog.Warn("teststream bytes - after LP", "value", string(v), "bytes", buf.Bytes())
 						return buf.Bytes()
+					}
+
+					// we need to change Coder and Pcollection in comps directly before they are used to build descriptors
+					for _, col := range t.GetOutputs() {
+						oCID := comps.Pcollections[col].CoderId
+						comps.Pcollections[col].CoderId = cID
+						slog.Debug("teststream: rewrite coder for output pcoll", "colId", col, "oldId", oCID, "newId", cID)
 					}
 				}
 
@@ -388,6 +406,7 @@ func executePipeline(ctx context.Context, wks map[string]*worker.W, j *jobservic
 				wk := wks[s.envID]
 				if err := s.Execute(ctx, j, wk, comps, em, rb); err != nil {
 					// Ensure we clean up on bundle failure
+					j.Logger.Error("Bundle Failed.", slog.Any("error", err))
 					em.FailBundle(rb)
 					return err
 				}
@@ -498,7 +517,7 @@ func buildTrigger(tpb *pipepb.Trigger) engine.Trigger {
 			Transforms: transforms,
 		}
 	case *pipepb.Trigger_AfterSynchronizedProcessingTime_:
-		panic(fmt.Sprintf("unsupported trigger: %v", prototext.Format(tpb)))
+		return &engine.TriggerAfterSynchronizedProcessingTime{}
 	default:
 		return &engine.TriggerDefault{}
 	}